# [Mono.Android] improve performance of new Java arrays