Context: https://github.com/dotnet/maui/blob/79f2fef9397bbfe2f6aa1527e40f933712a5d2ad/src/Core/src/Platform/Android/ColorStateListExtensions.cs#L50-L51

When running a `dotnet new maui` app, we noticed the log:

    W monodroid: typemap: failed to map managed type to Java type: System.Int32, System.Private.CoreLib, Version=6.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e (Module ID: 8e4cd939-3275-41c4-968d-d5a4376b35f5; Type token: 33554653)
    W monodroid-assembly: typemap: called from
    W monodroid-assembly: at Android.Runtime.JNIEnv.TypemapManagedToJava(Type )
    W monodroid-assembly: at Android.Runtime.JNIEnv.GetJniName(Type )
    W monodroid-assembly: at Android.Runtime.JNIEnv.FindClass(Type )
    W monodroid-assembly: at Android.Runtime.JNIEnv.NewArray(Array , Type )
    W monodroid-assembly: at Android.Runtime.JNIEnv.NewArray[Int32[]](Int32[][] )
    W monodroid-assembly: at Android.Content.Res.ColorStateList..ctor(Int32[][] , Int32[] )
    W monodroid-assembly: at Microsoft.Maui.Platform.ColorStateListExtensions.CreateButton(Int32 enabled, Int32 disabled, Int32 off, Int32 pressed)

We see this ~17 times on startup. In this case, the Android binding is:

    public unsafe ColorStateList (int[][]? states, int[]? colors) : base (IntPtr.Zero, JniHandleOwnership.DoNotTransfer)
    {
        //...
        IntPtr native_states = JNIEnv.NewArray (states);

Looking at the stack trace, it seems like this is happening for every
item in the array? We could probably call `FindClass (typeof (T))`
once, and use it for every item in the array. This would avoid many
typemap lookups.

Previously this showed up in `dotnet trace` output for a Pixel 5:

    6.05ms Mono.Android!Android.Runtime.JNIEnv.NewArray
    4.84ms Mono.Android!Android.Runtime.JNIEnv.NewArray(System.Array,System.Type)

After these changes:

    5.04ms Mono.Android!Android.Runtime.JNIEnv.NewArray
    3.83ms Mono.Android!Android.Runtime.JNIEnv.NewArray(System.Array,System.Type,intptr)

This change seems to save ~1ms on startup for `dotnet new maui` on a
Pixel 5 device.